perf(profile): AuthProvider + warm profile seed (#496)
Collapse dashboard auth/profile listeners into a single AuthProvider, gate profile getDoc on App Check, and seed the Profile form from auth userProfile when warm.

Closes #496
