+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "setlistpickem",
   "private": true,
-  "version": "1.20.16",
+  "version": "1.20.17",
   "type": "module",
   "scripts": {
     "dev": "vite",
