+++ src/main.jsx
@@ -6,6 +6,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import App from './app/App.jsx'
 import Ga4RouteListener from './app/Ga4RouteListener.jsx'
 import ScrollToTop from './app/ScrollToTop.jsx'
+import { AuthProvider } from './features/auth'
 import { initGa4 } from './shared/lib/ga4'
 import { initializeAppCheckDeferred } from './shared/lib/firebaseAppCheck'
 import { registerMessagingServiceWorker } from './shared/lib/firebaseMessaging'
@@ -33,9 +34,11 @@ root.render(
     <BrowserRouter future={{ v7_startTransition: true, v7_relativeSplatPath: true }}>
       <QueryClientProvider client={queryClient}>
         <HelmetProvider>
-          <ScrollToTop />
-          <Ga4RouteListener />
-          <App />
+          <AuthProvider>
+            <ScrollToTop />
+            <Ga4RouteListener />
+            <App />
+          </AuthProvider>
         </HelmetProvider>
       </QueryClientProvider>
     </BrowserRouter>
